+++ src/margin.rs
@@ -2,25 +2,25 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, Debug, Clone, Copy)]
 pub struct Margin {
-    pub top: u32,
-    pub bottom: u32,
-    pub left: u32,
-    pub right: u32,
+    pub top: f32,
+    pub bottom: f32,
+    pub left: f32,
+    pub right: f32,
 }
 
 impl Default for Margin {
     fn default() -> Self {
         Margin {
-            top: 0,
-            bottom: 0,
-            left: 0,
-            right: 0,
+            top: 0.0,
+            bottom: 0.0,
+            left: 0.0,
+            right: 0.0,
         }
     }
 }
 
 impl Margin {
-    pub fn new(top: u32, bottom: u32, left: u32, right: u32) -> Margin {
+    pub fn new(top: f32, bottom: f32, left: f32, right: f32) -> Margin {
         Margin {
             top,
             bottom,
@@ -31,19 +31,19 @@ impl Margin {
 }
 
 impl Margin {
-    pub fn with_top(self, top: u32) -> Margin {
+    pub fn with_top(self, top: f32) -> Margin {
         Margin { top, ..self }
     }
 
-    pub fn with_bottom(self, bottom: u32) -> Margin {
+    pub fn with_bottom(self, bottom: f32) -> Margin {
         Margin { bottom, ..self }
     }
 
-    pub fn with_left(self, left: u32) -> Margin {
+    pub fn with_left(self, left: f32) -> Margin {
         Margin { left, ..self }
     }
 
-    pub fn with_right(self, right: u32) -> Margin {
+    pub fn with_right(self, right: f32) -> Margin {
         Margin { right, ..self }
     }
 }
